requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.8.0rc2
Django==4.2.20
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-yasg==1.21.10
Faker==35.2.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==5.13.2
kombu==5.5.3
oauthlib==3.2.2
packaging==25.0
pillow==10.4.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
